Update to detect iOS
iOS security blocks AJAX downloads of PDF files via hidden frames like datafrm. By detecting if iOS is running and then changing the target, we can allow PDF files to download in a new tab, allowing direct downloading and reading or importation to iBooks or any other reader.

Another way might be a function to create a new _blank with a direct link to action.php, but this works.