Correct wrong use of correctDirectory, fixes #1072
If sendTorrent is called with a non-empty $directory parameter, it will always fail because correctDirectory returns a comparison with $topDirectory which is part of the app's URL, and not at all related for the filesystem. This commit fixes the problem so that if a $directory parameter is passed, the function will not fail.