Fix double URL encoding of torrent labels
`webui.js` encodes the label text before dispatching the web request, where the `setlabel()` function in the httprpc plugin`init.js` file also does a URL encode of the label text. The end result is the label looks and acts fine in the web-ui, but it is actually setting a URL encoded string of the label in rTorrent. This can cause unexpected behaviour.