Support intersecting selections when filtering torrents.
The State, Label, Tracker, and Search sections can now all have their
own value selected and the torrent list is filtered by the intersection
(AND) of all four sections.

Each section now has an "All" item at the top which is the default
selection.

When a new search is made all categories are set to "All" automatically
to preserve the behavior of search showing all search results.