Compress jquery.flot.js
There shouldn't be any need to ever change this wrapper. If we decide to update the version in the future, it can be recompressed and recommitted.

Install scripts are going to start shipping vmtouch improvements soon.  This reduces the largest file size from ~125KB to ~90KB. The memory filter can be set accordingly, to reduce memory consumption.