Integrate 'DarkBetter' & 'MaterialDesign' themes
This commit removes the DarkBetter and MaterialDesign themes from the project as submodules. It retains the commit history for compatibility with third-party repositories.

It adds the DarkBetter and MaterialDesign themes directly to the main project instead. This is required because GitHub developers have chosen not to properly support submodules.

Submodules are great for developers but terrible for end users.

When users click the 'Download Zip' button on the GitHub repository, the submodules do not download. This can create breakage when upgrading the ruTorrent version.

Furthermore, our repository releases do not draft with the submodules included. It would be required to add and maintain third-party code to automate this process. It is not feasible to draft releases manually without the risk of human error.