Remove case insensitivity from PHP named-constants
This feature is deprecated as of PHP 7. It will produce a warning for PHP 8 if set to true. We should not be using this feature.

It allows inconsistent naming and usage of PHP constants. It suppresses the notice and makes things work properly.

It may introduce regressions and make them harder to find. We need constants and variables to be easily distinguishable to review code. We also require feedback from PHP to properly test code.