graph: Improve speed and trafic graphs
Resolved issue: Changing maxSeconds was not working correctly.
Furthermore, both jquery.fplot.js graphs for speed and
trafic were recreating the plot for every update.

Create new class which uses plot.draw() to update

Allow mouse dragging and scrolling to pan and zoom the speed graph

Trigger onShow correctly if the activeView is added by a plugin

Move the tooltip of trafic graph above the bar
and the tooltip of speed graph above the cursor