webui: Run addTorrents as a background task
Resovled issue: when updating thousands of torrents
the UI gets unresponsive due to a busy event loop.

Add a background task class which uses `requestIdleCallback`

Add fallback for browsers not supporting `requestIdleCallback`

Start a BackgroundTask for addTorrents and cancel previous task.
After the task is done we update the DOM via `requestAnimationFrame`.