WhichInstance: Always prune cache on first initialization
If the user uninstalls a package on their operating system and sets $do_diagnostic to false, it will throw an error. It's safer to always prune cache once when the WhichInstance is first initialized. It doesn't take much time for PHP to check if the binary exists and is executable. The results are also cached properly from PHP.
https://www.php.net/manual/en/function.is-executable.php