Fix adding new torrents with reverse proxy
Closes #2526. Improves 345dbe5.

This commit fixes adding new torrents when a reverse proxy is enabled. It also adds support for subdirectories in front of rutorrent on reverse proxies.

The solution is not to reference variables which aren't set, when building the path to addtorrent.php. Also, to reference the php script with dot forward slash, to say it's located inside the current folder.

ruTorrent does not support both cross origin and reverse proxy for adding torrents at this moment in time. Please pick one. All this commit does is avoid triggering a secondary error, when attempting to support cross origin.