tree-view: Rewrite updateCustomLabels
Now, the custom labels are explicitly processed as a tree.
Empty label nodes which only have one child are flattened
unless `webui.show_empty_path_labels` is `true`.
Additionally, the children of `clabel_cont` and `lbll` now
both reside in `plabel_cont`
and are updated in `updateCustomLabels`, as well.

Fixes #2548 #2549 #2550