Drop support for RPC & optimize for httprpc
We can no longer support RPC for stability reasons. The web interface may start to fail at just 150 torrents.

The newer httprpc alternative is now a lot more stable and polished. By dropping support for rpc, we can bring the web interface update settings into this decade.

It's required to reduce the request time out, to ensure we do not overload  web servers. We can also run a lower update interval, due to threaded web interface updates in the browser.

This does not impact critical requests with no time out. If the request can potentially take longer than 10s, we should not be using a timeout for it.