Increase height of torrent creator file browser
This commit increases the height of the file browser i-frame for the torrent creator to 300 pixels. It makes finding files or folders much easier.

It also introduces a feature where the height of the rDirBrowser i-frame can be customized for each instance. An optional parameter is added to the initializer.