Convert	recentTrackers->last_used to an	array
Previously, we were storing a string, and if a user created 20 torrents
in a row, it would add 20 trailing carriage returns (\r) to the	box.

We already do the logic	of stripping empty lines for other purposes,
lets reuse that here.

Add de-duplication of last used trackers (to match the list array).

Also eliminate $announce_list, because it's never used, and `grep -rn announce_list * | grep global`
indicates no other function is using it without it being passed.