Remove tr index attribute
It looks like it has lost it's original purpose and is now identical
on every single torrent when you first load it. It does look like
it will increment as new torrents are adedd. But I don't see a lot
of utliity. Nobody needs to reference the 57th torrent added to the
client.

It's also incurring a hefty performance penalty in the code. Removing
it now makes syncDom 5x faster. From 270 ms on my machine to 57 ms
reliabily in Chrome.