Fix Chrome performance regression (#2960)
* Fix Chrome performance regression

Now that dxSTable.rows is a getter, and not a variable, we need to
be mindful about using it in for loops. On an ruTorrent with 20K
torrents, putting it into the for loop, instead of copying it into
a variable will result in 20K executions of the getter function
and tank performance.

By checking the rowcount once, before we start the loop, we can
keep the getter (reducing the risk of OOS bugs), without suffering
the performance penalty.

Related to #2830

* Remove tr index attribute

It looks like it has lost it's original purpose and is now identical
on every single torrent when you first load it. It does look like
it will increment as new torrents are adedd. But I don't see a lot
of utliity. Nobody needs to reference the 57th torrent added to the
client.

It's also incurring a hefty performance penalty in the code. Removing
it now makes syncDom 5x faster. From 270 ms on my machine to 57 ms
reliabily in Chrome.